requirements.txt
requests==2.22.0
psycopg2==2.8.4
Django==2.2.7
djangorestframework==3.10.3
django-filter==2.2.0
celery==4.3.0
redis==3.3.11
